feat: 상품의 position과 CartRepository::deleteCartProduct의 id가 동일한 값임을 보장할 수 없다는 문제를 해결

+++ app/src/main/java/woowacourse/shopping/data/DefaultCartRepository.kt
@@ -13,5 +13,5 @@ class DefaultCartRepository(
 
     override suspend fun getAllCartProducts(): List<CartUiModel> = dao.getAll().map { it.toPresentation() }
 
-    override suspend fun deleteCartProduct(id: Int) = dao.delete(id.toLong())
+    override suspend fun deleteCartProduct(id: Long) = dao.delete(id)
 }
